agents-lab/rag-from-scratch/requirements.txt
numpy
sentence-transformers
openai
tqdm
python-dotenv
